requirements.txt
django==3.2.16
django-environ==0.9.0
django-cors-headers==3.13.0
djangorestframework==3.14.0

argon2-cffi==21.3.0
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
PyJWT==2.5.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0

Pillow==9.3.0
django-storages==1.13.1
django-ckeditor==6.7.1
psycopg2-binary==2.9.9
python-slugify==6.1.2

django-import-export==3.3.0
django-guardian==2.4.0

django-unfold==0.21.1
django-ckeditor-5==0.2.12

